+++ testdata/codegen_matrix/fixtures/immutable_defaults/src/ent/user.ts
@@ -1,0 +1,8 @@
+import { AlwaysAllowPrivacyPolicy } from "@snowtop/ent";
+import { UserBase } from "./generated/user_base";
+
+export class User extends UserBase {
+  getPrivacyPolicy() {
+    return AlwaysAllowPrivacyPolicy;
+  }
+}